Adding jump
Problems with the jump it was basically saying if mollychar.y is
greater than or equal to the stage height then jump and land back on
the stage height, the stage height is the bottom border of the canvas
that she's in I think, so she was basically falling down to the bottom
of the canvas. So instead of saying jump if she's greater than or equal
to the stage height just have her jump if her y axis is greater than or
equal to her x axis and land back on her x axis.